refactor(restaurants-menus): update import statements to include file extensions
- Modified import statements across all components to include the '.js' file extension for consistency.
- Enhanced code clarity and maintainability by adhering to established import conventions.
- Ensured all components are aligned with the updated file structure.
