+++ packages/headless-components/restaurants-menus/src/react/index.tsx
@@ -1,8 +1,8 @@
-export * as Menus from './Menus';
-export * as Menu from './Menu';
-export * as Section from './Section';
-export * as Item from './Item';
-export * as Variant from './Variant';
-export * as Label from './Label';
-export * as ModifierGroup from './ModifierGroup';
-export * as Modifier from './Modifier';
+export * as Menus from './Menus.js';
+export * as Menu from './Menu.js';
+export * as Section from './Section.js';
+export * as Item from './Item.js';
+export * as Variant from './Variant.js';
+export * as Label from './Label.js';
+export * as ModifierGroup from './ModifierGroup.js';
+export * as Modifier from './Modifier.js';
