+++ packages/headless-components/restaurants-menus/src/services/index.ts
@@ -4,4 +4,4 @@ export {
   loadMenusServiceConfig,
   type MenusServiceConfig,
   type MenusServiceAPI,
-} from './menus-service';
+} from './menus-service.js';
